REevo: phase 1: return

+++ scripts/rm_failed_train_log.py
@@ -5,8 +5,9 @@ import sys
 
 tmp_file = './tmp/tmp.txt'
 
+if len(sys.argv) < 4:
+    print(f'{sys.argv[0]} [log_dir] [retain TAG] [REMOVE(0/1)]')
 if len(sys.argv) < 3:
-    print(f'{sys.argv[0]} [log_dir] [retain TAG]')
     exit()
 
 dis_dir = sys.argv[1]
@@ -20,13 +21,19 @@ lines = [x for x in open(tmp_file, 'r').readlines() if '.out' in x]
 junk = []
 
 for l in lines:
-    tks = l.strip().split(' ')
-    if tks[3] != tag:
+    tks = l.strip().split(' ')#
+    #if tks[3] != tag:#
+    if tag not in l:
         junk += [f'{dis_dir}/{tks[-1]}']
         junk += [f'{dis_dir}/{tks[-1]}'[:-3]+'err']
 
 print(f'Failed training jobs: {int(len(junk)/2)}')
 
-for j in junk:
-    print(f'removing {j}')
-    os.remove(j)
+if len(sys.argv) > 3 and int(sys.argv[3]) == 1:
+    for j in junk:
+        print(f'removing {j}')
+        os.remove(j)
+
+    print('Failed logs removed.')
+else:
+    print('Use Flag {1} to remove failed logs.')
